+++ src/Credfeto.Dispatcher.GitHub.DataTypes/PullRequestDetails.cs
@@ -19,6 +19,5 @@ public sealed record PullRequestDetails(
     IReadOnlyList<LinkedItem> LinkedItems,
     ItemRepository Repository,
     LastNotification LastNotification,
-    string? Author,
-    string? HeadBranchName
+    string? Author
 );